Finished with IM1
Finished debugging using IM1.  Created a TopModuleSC to debug with IM1. The relevant changes I made to TopModuleSC were also made to TopModuleV2. Essentially I added a wire that handles making sure that there is no writing done to the register if we have a madd/msub/etc. instructions with the exceptions being the mfhi/mflo instructions. I also edited my top_tb to use the topmodulesc, so watch out for that. Lastly, I edited the hl reg because the output was registering a cycle after the  execution stage (so in the MEM stage) so the processor was just writing a 32-bit of X's because of it. Now the output of the HL Reg is registered at the EX stage, so the wb_data is correct